refactor socketHandler to use socketFunctions.js
